basic read support for linux !
