+++ README.md
@@ -1,5 +1,8 @@
 # memorpy
-Python library using ctypes to search/edit windows programs memory. Some functions are not implemented as I only implemented what I needed at the time I wrote this ;) But the basic functions should works just fine
+Python library using ctypes to search/edit windows and linux programs memory. Some functions are not implemented as I only implemented what I needed at the time I wrote this ;) But the basic functions should works just fine
+
+#TODO
+implement LinProcess.write method on linux
 
 #usage examples :
 In this example open a notepad.exe (x32) and type in some text we will edit from memory !
