+++ memorpy/LinStructures.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: UTF8 -*-
+
+#Use some Windows constants for compatibility
+PAGE_EXECUTE_READWRITE = 64
+PAGE_EXECUTE_READ = 32
+PAGE_READONLY = 2
+PAGE_READWRITE = 4
